+++ tfexec/e2etest/testdata/basic/terraform.tfstate
@@ -1,0 +1,26 @@
+{
+  "version": 4,
+  "terraform_version": "0.12.24",
+  "serial": 1,
+  "lineage": "3d011417-36e1-8302-77c5-7e45fdf14235",
+  "outputs": {},
+  "resources": [
+    {
+      "mode": "managed",
+      "type": "null_resource",
+      "name": "foo",
+      "provider": "provider.null",
+      "instances": [
+        {
+          "schema_version": 0,
+          "attributes": {
+            "id": "5510719323588825107",
+            "triggers": null
+          },
+          "private": "bnVsbA=="
+        }
+      ]
+    }
+  ]
+}
+
